feat: update todo test

+++ app/Services/Impl/TodoListServiceImpl.php
@@ -28,6 +28,8 @@ class TodoListServiceImpl implements TodoListService
     {
         $todo = Todo::query()->find($todoId);
 
-        $todo->delete();
+        if ($todo !== null) {
+            $todo->delete();
+        }
     }
 }
